Remove unnecessary rx frequency check
The receive frequency doesn't need to be above the base frequency. Consider the standard 900 MHz band plan, `0,935006250,6.25,-39.000,12.5,`; base transmit frequencies are always in 935-940 MHz, and base receive frequencies are always in 896-901 MHz. The range 901-935 MHz is outside the scope of the band plan, and it is impossible to cover the entire 896-940 MHz range in 6.25 kHz steps. Instead, the transmit frequency is sent over the air and the receiver calculates the receive frequency, so it is out of the scope of DVMHost to ensure the receive frequency is within the band plan.